progress on the transparent feature scaling problem.
